moved objects to abstraction
